+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Don't explicitly use internal igraph dependencies.
 - Update to igraph 0.10.12
+- Explicitly include needed igraph headers instead of catchall "igraph.h"
 
 ## [v0.1.3] 2024-04-19